Update 'create_restriction_cadastre_wizard' wizard
