+++ asistente_ladm_col/asistente_ladm_col_plugin.py
@@ -76,7 +76,7 @@ from .gui.create_plot_cadastre_wizard import CreatePlotCadastreWizard
 from .gui.wizards.cadastre.surveying.create_points_cadastre_wizard import CreatePointsCadastreWizard
 from .gui.create_property_record_card_prc import CreatePropertyRecordCardPRCWizard
 from .gui.create_responsibility_cadastre_wizard import CreateResponsibilityCadastreWizard
-from .gui.create_restriction_cadastre_wizard import CreateRestrictionCadastreWizard
+from .gui.wizards.cadastre.rrr.create_restriction_cadastre_wizard import CreateRestrictionCadastreWizard
 from .gui.wizards.cadastre.rrr.create_right_cadastre_wizard import CreateRightCadastreWizard
 from .gui.create_spatial_source_cadastre_wizard import CreateSpatialSourceCadastreWizard
 from .gui.create_parcel_valuation_wizard import CreateParcelValuationWizard
